+++ changelog.md
@@ -5,6 +5,7 @@ Changelog
 
 - Fixed obd.dump putting hashes in the wrong places when recording messages that don't match selected protocol. (Previously would cause 'fromhex' error on obd.play)
 - Fixed issue where selecting adaptive timing causes indentation error
+- Fixed command 'minionutil.change_master' to only perform changes directly in the minion config file.
 
 ## v1.22